+++ ios/Geoluciole/Geoluciole/Models/Badge.swift
@@ -17,7 +17,7 @@ class Badge {
     var latitude: Double?
     var longitude: Double?
     var proximity: Int?
-    var distance: Int?
+    var distance: Double?
     var resource: String
     var isObtain: Bool
 
@@ -29,7 +29,7 @@ class Badge {
         self.latitude = dictionnary["latitude"] as? Double ?? nil
         self.longitude = dictionnary["longitude"] as? Double ?? nil
         self.proximity = dictionnary["proximity"] as? Int ?? nil
-        self.distance = dictionnary["distance"] as? Int ?? nil
+        self.distance = dictionnary["distance"] as? Double ?? nil
         self.resource = dictionnary["resource"] as! String
         self.isObtain = dictionnary["is_obtain"] as? Bool ?? false
     }
